Move MetricType into correct package
