+++ m3/src/main/java/com/uber/m3/tally/m3/MetricType.java
@@ -18,7 +18,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-package com.uber.m3.tally;
+package com.uber.m3.tally.m3;
 
 /**
  * Enum of the type of metric.
